+++ packages/sdk/client-services/src/packlets/services/index.ts
@@ -2,7 +2,6 @@
 // Copyright 2022 DXOS.org
 //
 
-export * from './local-client-services';
 export * from './service-context';
 export * from './service-host';
 export * from './service-registry';
