+++ packages/sdk/client/eslint-config.json
@@ -5,9 +5,23 @@
   // This is only required in order to do packlet linting, otherwise rules will
   //   automatically be applied from the root config.
   "extends": [
-    "../../../.eslintrc.packlets.js"
+    "../../../.eslintrc.js"
   ],
   "parserOptions": {
     "project": "./packages/sdk/client/tsconfig.json"
-  }
+  },
+  "overrides": [
+    {
+      "files": "**/!(*.test).{ts,tsx}",
+      "rules": {
+        "@typescript-eslint/no-restricted-imports": ["error", {
+          "paths": [{
+            "name": "@dxos/client-services",
+            "message": "Importing whole client-services package can be slow. Please mark the import as type-only or use a dynamic import where needed.",
+            "allowTypeImports": true
+          }]
+        }]
+      }
+    }
+  ]
 }
